Import template with context (#1)
* remove f string

* import template no context

* use context

* import directory

* readme WIP

* readme WIP

* readme WIP

* readme WIP

* readme WIP

* app improvements

* change poster

+++ src/main.py
@@ -1,7 +1,5 @@
 import os
-from pathlib import Path
 from dotenv import load_dotenv
-import requests
 import supervisely as sly
 
 # load ENV variables for debug, has no effect in production
@@ -10,46 +8,34 @@ load_dotenv(os.path.expanduser("~/supervisely.env"))
 
 
 class MyImport(sly.app.Import):
-    def process(self, workspace_id: int, path: str, is_directory: bool):
+    def process(self, context: sly.app.Import.Context):
         # create api object to communicate with Supervisely Server
         api = sly.Api.from_env()
 
-        # create project with the same name as input file and empty dataset in it
-        project_name = Path(path).stem
-        project = api.project.get_or_create(workspace_id, project_name)
-        print(f"Working project: id={project.id}, name={project.name}")
-        dataset = api.dataset.create(project.id, "dataset", change_name_if_conflict=True)
-        print(f"Created dataset: id={dataset.id}, name={dataset.name}")
+        # list images in directory
+        images_names = []
+        images_paths = []
+        for file in os.listdir(context.path):
+            file_path = os.path.join(context.path, file)
+            images_names.append(file)
+            images_paths.append(file_path)
 
-        # read input file, remove empty lines + leading & trailing whitespaces
-        with open(path) as file:
-            lines = [line.strip() for line in file.readlines() if line.strip()]
-
-        # process text file and remove empty lines
-        progress = sly.Progress("Processing urls", total_cnt=len(lines))
-        for index, img_url in enumerate(lines):
+        # process images and upload them by paths
+        progress = sly.Progress("Processing images", total_cnt=len(images_names))
+        for img_name, img_path in zip(images_names, images_paths):
             try:
-                img_ext = Path(img_url).suffix
-                img_name = f"{index:03d}{img_ext}"
-                img_path = os.path.join(os.getcwd(), "data", img_name)
-
-                # download image
-                response = requests.get(img_url)
-                with open(img_path, "wb") as file:
-                    file.write(response.content)
-
-                # upload image into dataset on Supervisely server
-                info = api.image.upload_path(dataset.id, img_name, img_path)
-                sly.logger.trace(f"Image has been uploaded: id={info.id}, name={info.name}")
+                # upload image by path
+                api.image.upload_path(dataset_id=context.dataset_id, name=img_name, path=img_path)
 
                 # remove local file after upload
-                os.remove(img_path)
+                if sly.utils.is_production():
+                    os.remove(img_path)
             except Exception as e:
-                sly.logger.warn(f"Skip image", extra={"url": img_url, "reason": repr(e)})
+                sly.logger.warn("Skip image", extra={"name": img_name, "reason": repr(e)})
             finally:
                 progress.iter_done_report()
 
-        return project.id
+        return context.project_id
 
 
 app = MyImport()
